requirements.txt
requests
dotenv
pandas
minio
lxml
dbt-snowflake
snowflake-connector-python[pandas]
dbt-core
apache-airflow[cncf.kubernetes,celery,postgres,google,amazon,s3,http,slack,statsd,jdbc,ssh]==2.9.2 # Choose specific extras based on your needs, a smaller set is good for dev
apache-airflow-providers-cncf-kubernetes # If using KubernetesExecutor, ensure it's here
apache-airflow-providers-postgres
pyarrow
psycopg2-binary
black
pytest
Flask-Limiter==3.3.1
